+++ Capturis-Client/src/app/assessment-results/assessment-results.component.css
@@ -16,8 +16,6 @@
 }
 
 .example-card {
-  width: 100%;
-  height: 100%;
   position: absolute;
   left: 20%;
   top: 30%;
