+++ Capturis-Client/src/app/login/login.component.css
@@ -1,86 +1,3 @@
-
-/*
-html, body{
-
-	background-color: WHITE;
-
-}
-
-form {
-    border: 5px solid #4CAF50;
-	background-color:white;
-	width: 100%;
-}
-
-.midrow{
-	margin-top: 10%;
-	height: 50%;
-	width: 100%;
-
-}
-
-input[type=text], input[type=password] {
-    width: 60%;
-	text-align: center;
-    padding: 12px 20px;
-    margin: 0 auto;
-    display: inline;
-    border: 1px solid #ccc;
-    box-sizing: border-box;
-}
-
-button {
-    background-color: #4CAF50;
-    color: white;
-    padding: 14px 20px;
-	display: block;
-    margin: 0 auto;
-    border: none;
-    cursor: pointer;
-    width: 50%;
-}
-
-
-button:hover {
-    opacity: 0.8;
-}
-
-.imgcontainer {
-    text-align: center;
-    margin: 24px 0 12px 0;
-	background-color: #4CAF50;
-}
-
-img.logo {
-    width: 100%;
-    border-radius: 0%;
-}
-
-.container {
-    padding: 16px;
-	vertical-align: middle;
-	text-align: center;
-	background-color: white;
-	border-style: solid;
-	border-width: 50px;
-	border-color: #4CAF50;
-	width: 550px;
-	height: 100%;
-}
-
-span.psw {
-    float: right;
-    padding-top: 16px;
-    vertical-align: middle;
-}
-label {
-  display: block;
-  width: 140px;
-  padding: 0px 20px;
-  text-align: center;
-}
-*/
-
 .container {
 	margin-top: 20px;
 }
