+++ Capturis-Client/src/app/models/assessment-results.model.ts
