+++ Capturis-Client/src/app/models/assessment.model.ts
