+++ Capturis-Client/src/app/models/manager-dashboard.model.ts
