+++ Capturis-Client/src/app/user-dashboard/user-dashboard.component.css
@@ -1,106 +1,107 @@
-      h2 {
-      text-align: center;
-      padding:10px;
-      }
-      .box
-      {
-      background-color: #42683c;
-      width: 100%;
-      min-width: 250px;
-      height: 50px;
-      padding: 30px;
-      margin: 3px auto;
-      }
-      .box2
-      {
-      border:lightblue 1px solid;
-      width: 100%;
-      min-width: 250px;
-      height: 250px;
-      padding: 20px;
-      margin: 1px auto;
-      }
-      .rightBox
-      {
-        width: 45%;
-        float: right;
-      }
-      .leftBox
-      {
-        width: 55%;
-        float: left
-      }
-      .minutes
-      {
-        margin: auto;
-        width: 100%;
-        height: 100%;
-        padding: 10px;
-      }
-      .box3
-      {
-      background-color: #d3d3d3;
-      border:blue 1px solid;
-      height: 200px;
-        margin: auto;
-        width: 75%;
-        padding: 10px;
-      }
-      .desctext
-      {
-      color: #ffffff;
-      position: relative;
-      top: 50%;
-      transform: translateY(-50%);
-      font: 500 20px/32px Roboto, 'Helvetica Neue', sans-serif;
-      }
-      .userDashboard
-      {
-        width: 100%;
-        height: 100%;
-        white-space: nowrap;
-        overflow: auto;
-        display: inline-block;
-      }
-      .AssessmentDesc
-      {
-      padding-left: 8px;
-      height: 350px;
-      width: calc(25% - 6px);
-      min-width: 250px;
-      display: inline-block;
-      }
-      .assessmentButton
-      {
-      background-color: #4CAF50;
-      width: 100%;
-      height: 35px;
-      float: right;
-      color: #ffffff;
-      margin-bottom: 20px;
-      }
-      .HistoryBox
-      {
-        display: block;
-      background-color: #42683c;
-      height: 60px;
-        padding-left: 24px;
+h2 {
+  text-align: center;
+  padding: 10px;
+}
 
-        width: calc(100% - 12px);
-        position: absolute;
-        left: 50%;
-        transform: translateX(-50%);
-      }
-      .historyTable
-      {
-        margin-top: 90px;
-        width: calc(100% - 12px);
-        position: absolute;
-        left: 50%;
-        transform: translateX(-50%);
-      }
-      .logout {
-      	font-size: 30px;
-      	float:right;
-      	margin-right: 100px;
-      }
+.box {
+  background-color: #42683c;
+  width: 100%;
+  min-width: 250px;
+  height: 50px;
+  padding: 30px;
+  margin: 3px auto;
+}
+
+.box2 {
+  border: lightblue 1px solid;
+  width: 100%;
+  min-width: 250px;
+  height: 250px;
+  padding: 20px;
+  margin: 1px auto;
+}
+
+.rightBox {
+  width: 45%;
+  float: right;
+}
+
+.leftBox {
+  width: 55%;
+  float: left
+}
+
+.minutes {
+  margin: auto;
+  width: 100%;
+  height: 100%;
+  padding: 10px;
+}
+
+.box3 {
+  background-color: #d3d3d3;
+  border: blue 1px solid;
+  height: 200px;
+  margin: auto;
+  width: 75%;
+  padding: 10px;
+}
+
+.desctext {
+  color: #ffffff;
+  position: relative;
+  top: 50%;
+  transform: translateY(-50%);
+  font: 500 20px/32px Roboto, 'Helvetica Neue', sans-serif;
+}
+
+.userDashboard {
+  width: 100%;
+  height: 100%;
+  white-space: nowrap;
+  overflow: auto;
+  display: inline-block;
+}
+
+.AssessmentDesc {
+  padding-left: 8px;
+  height: 350px;
+  width: calc(25% - 6px);
+  min-width: 250px;
+  display: inline-block;
+}
+
+.assessmentButton {
+  background-color: #4CAF50;
+  width: 100%;
+  height: 35px;
+  float: right;
+  color: #ffffff;
+  margin-bottom: 20px;
+}
+
+.HistoryBox {
+  display: block;
+  background-color: #42683c;
+  height: 60px;
+  padding-left: 24px;
+
+  width: calc(100% - 12px);
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+}
+
+.historyTable {
+  margin-top: 90px;
+  width: calc(100% - 12px);
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+}
+
+.logout {
+  font-size: 30px;
+  float: right;
+  margin-right: 100px;
+}
